% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_sat.R
\name{ar_sat}
\alias{ar_sat}
\title{Calculate argon saturation concentration}
\usage{
ar_sat(t, sal, baromet.press.atm)
}
\arguments{
\item{t}{water temperature in degrees Celsius}

\item{sal}{salinity in per mille (ppt)}

\item{baromet.press.atm}{barometric pressure in atmospheres}
}
\value{
equilibrium dissolved Ar concentration in micromoles per
  kilogram (umol/kg)
}
\description{
Computes the equilibrium dissolved argon (Ar) concentration in water at a
given temperature, salinity, and barometric pressure using the equation of
Hamme and Emerson (2004, Eqns. 1-2, Table 4). Vapor pressure of water is
calculated via the Antoine equation (Stull 1947) and used to apply a
pressure correction to the saturation concentration.
}
\examples{
# Ar saturation at 20 degrees C, fresh water, 1 atm
ar_sat(t = 20, sal = 0, baromet.press.atm = 1)

}
\references{
Hamme, R. C. & Emerson, S. R. (2004). \emph{The solubility of neon,
nitrogen and argon in distilled water and seawater}, Deep-Sea Research I,
51(11), 1517-1528.

Stull, D. R. (1947). \emph{Vapor Pressure of Pure Substances. Organic and
Inorganic Compounds.} Industrial & Engineering Chemistry, 39(4), 517-540.
doi: 10.1021/ie50448a022
}
