% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molar_to_mass.R
\name{molar_to_mass}
\alias{molar_to_mass}
\title{Calculate mass concentrations in mg/L given molar concentrations in umol/L}
\usage{
molar_to_mass(species, value, inUnits = "umol/L", outUnits = "mg/L")
}
\arguments{
\item{species}{Molecule, must be "N2", "O2", or "Ar"}

\item{value}{Concentration of molecule, numeric value}

\item{inUnits}{Units of molecule concentration, default is "umol/L"}

\item{outUnits}{Desired units for output, must be "mg/L", "g/L", or "ug/L". default is "mg/L"}
}
\value{
A numeric vector
}
\description{
Calculate mass concentrations in mg/L given molar concentrations in umol/L
}
\examples{
molar_to_mass(species = "N2", value = 12.1, inUnits = "umol/L", outUnits = "mg/L")
}
