% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n2sat.R
\name{n2sat}
\alias{n2sat}
\title{Calculate dinitrogen gas concentrations at saturation given water temperature, atmospheric pressure, and salinity}
\usage{
n2sat(temp, pressure, pressUnits, salinity = 0, outUnits = "mg/L")
}
\arguments{
\item{temp}{Water temperature in degrees C}

\item{pressure}{Barometric pressure, where `pressUnits` denotes units}

\item{pressUnits}{Units of barometric pressure, must be "atm", "hPa", "Torr", "psi", "bar", "inHg", or "mmHg"}

\item{salinity}{Water salinity in units of per mille. Default is 0.}

\item{outUnits}{Preferred units for output, must be "mg/L" or "umol/L"}
}
\value{
A numeric vector
}
\description{
Calculate dinitrogen gas concentrations at saturation given water temperature, atmospheric pressure, and salinity
}
\examples{
n2sat(temp = 25, pressure = 29.8, pressUnits = "inHg", outUnits = "mg/L")
}
\references{
Hamme, R. C., & Emerson, S. R. (2004). The solubility of neon, nitrogen and argon in distilled water and seawater. Deep Sea Research Part I: Oceanographic Research Papers, 51(11), 1517–1528. https://doi.org/10.1016/j.dsr.2004.06.009
}
