% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2_sat.R
\name{o2_sat}
\alias{o2_sat}
\title{Calculate oxygen saturation concentration}
\usage{
o2_sat(t, sal, baromet.press.atm)
}
\arguments{
\item{t}{water temperature in degrees Celsius}

\item{sal}{salinity in per mille (ppt)}

\item{baromet.press.atm}{barometric pressure in atmospheres}
}
\value{
equilibrium dissolved oxygen concentration in micromoles per
  kilogram (umol/kg)
}
\description{
Computes the equilibrium dissolved oxygen concentration in water at a given
temperature, salinity, and barometric pressure using the combined fit
equation of Garcia and Gordon (1992, Eqn. 8). Vapor pressure of water is
calculated via the Antoine equation (Stull 1947) and used to apply a
pressure correction to the saturation concentration.
}
\examples{
# O2 saturation at 20 degrees C, fresh water, 1 atm
o2_sat(t = 20, sal = 0, baromet.press.atm = 1)

}
\references{
Garcia, H., and L. Gordon (1992), \emph{Oxygen solubility in seawater:
Better fitting equations}, Limnology and Oceanography, 37(6).

Stull, D. R. (1947). \emph{Vapor Pressure of Pure Substances. Organic and
Inorganic Compounds.} Industrial & Engineering Chemistry, 39(4), 517-540.
doi: 10.1021/ie50448a022
}
