% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2sat.R
\name{o2sat}
\alias{o2sat}
\title{Calculate dissolved oxygen concentrations at saturation given water temperature, atmospheric pressure, and salinity}
\usage{
o2sat(temp, pressure, pressUnits, salinity = 0, outUnits)
}
\arguments{
\item{temp}{Water temperature in degrees C}

\item{pressure}{Barometric pressure, where `pressUnits` denotes units}

\item{pressUnits}{Units of barometric pressure, must be "atm", "hPa", "Torr", "psi", "bar", "inHg", or "mmHg"}

\item{salinity}{Water salinity in units of per mille. Default is 0.}

\item{outUnits}{Preferred units for output, must be "mg/L" or "umol/L"}
}
\value{
A numeric vector
}
\description{
Calculate dissolved oxygen concentrations at saturation given water temperature, atmospheric pressure, and salinity
}
\examples{
o2sat(temp = 25, pressure = 29.8, pressUnits = "inHg", outUnits = "mg/L")
}
\references{
Garcia, H. E., & Gordon, L. I. (1992). Oxygen solubility in seawater: Better fitting equations. Limnology and Oceanography, 37(6), 1307–1312. https://doi.org/10.4319/lo.1992.37.6.1307
}
