% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressureCorrection.R
\name{pressureCorrection}
\alias{pressureCorrection}
\title{Correct barometric pressure for vapor pressure of water}
\usage{
pressureCorrection(temp, barpress_atm)
}
\arguments{
\item{temp}{Water temperature in degrees C}

\item{barpress_atm}{Barometric pressure in atm}
}
\value{
Barometric pressure corrected for water vapor pressure
}
\description{
Correct barometric pressure for vapor pressure of water
}
