% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimsy.R
\name{mimsy}
\alias{mimsy}
\title{\code{mimsy} Calculate dissolved gas concentration from raw MIMS signal data}
\usage{
mimsy(filename, bgcorr = FALSE, barpress, tz = Sys.timezone(),
  barpress_units, salinity = 0, std.temps)
}
\arguments{
\item{filename}{the name of the .csv file containing raw MIMS data. See github readme for .csv formatting guide}

\item{bgcorr}{a logical value indicating whether a background correction should be performed. Defaults to FALSE. If TRUE, user must specify values to use for the background correction.}

\item{barpress}{numeric, the barometric pressure of the room while samples were run on the MIMS. User may input a vector, script will take mean value.}

\item{tz}{a character string that specifies which time zone to parse the date with. Defaults to the user's system time zone setting. The string must be a time zone that is recognized by the user's OS. See \code{\link[lubridate]{lubridate::ymd_hms}} for more information.}

\item{barpress_units}{character string indicating the units of barometric pressure. Must be one of "atm", "hPa", "psi", "bar", or "Torr"}

\item{salinity}{numeric, the salinity of standards, in units of per mille. Defaults to 0.}

\item{std.temps}{a numeric vector (maximum length 2) of the water temperatures (degC) of the standard baths.}
}
\value{
list, $results containing the calculated dissolved gas concentrations in units of microM and mg of samples, $solubility.Concentrations containing the caluated solubility of gas at standard temperatures, $calibration.Factors containing the calibration factors for each standard group at each standard temperature, $calibration.DriftCorrection containing the drift corrected slope and intercept values for each gas, and $results.full containing all provided and calculated data, for both standards and samples
}
\description{
Returns a list containing a user-friendly dataframe of dissolved gas concentrations (uM and mg) of samples, a dataframe of the calculated gas solubility concentrations at standard temperatures, a dataframe of the calibration factors, a dataframe of the drift correction factors used, and a dataframe of the full results, including interpolated calibration slopes and calculated concentrations of standards.
}
\examples{
mydata <- mimsy(filename = './RawData/MIMS_data.csv', barpress = 981.2, barpress_units = 'hPa', std.temps = c(12.5, 15.2))

}
\references{
Garcia, H., and L. Gordon (1992), \emph{Oxygen solubility in seawater: Better fitting
equations}, Limnology and Oceanography, 37(6).

Benson, B. B. & Krause, D. (1984). \emph{The concentration and isotopic
fractionation of oxygen dissolved in freshwater and seawater in equilibrium
with the atmosphere.} Limnology and Oceanography, 29(3), 620-632.
doi:10.4319/lo.1984.29.3.0620

Stull, D. R. (1947). \emph{Vapor Pressure of Pure Substances. Organic and
Inorganic Compounds.} Industrial & Engineering Chemistry, 39(4), 517-540.
doi: 10.1021/ie50448a022

Hamme, R. C. & Emerson, S. R. (2004). \emph{The solubility of neon, nitrogen and argon
in distilled water and seawater}, Deep-Sea Research I, 51(11), 1517-1528.
}
