% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimsy.R
\name{mimsy}
\alias{mimsy}
\title{\code{mimsy} functionshortdescription}
\usage{
mimsy(filename, bgcorr = FALSE, barpress_units, salinity = 0,
  twoPoint = TRUE, std.temp)
}
\arguments{
\item{filename}{the name of the .csv file containing raw MIMS data. See github readme for .csv formatting guide}

\item{bgcorr}{a logical value indicating whether a background correction should be performed. Defaults to FALSE. If TRUE, user must specify values to use for the background correction.}

\item{barpress_units}{units of measured atmospheric pressure. function recognizes "atm", "hPa", (more units coming with update...)}

\item{salinity}{salinity of standards, in units of per mille. defaults to 0.}

\item{twoPoint}{a logical value indicating whether a two point temperature calibration was used (TRUE) or a single point temperature calibration was used (FALSE). Defaults to TRUE.}

\item{std.temp}{a numberic vector (maximum length 2) of the water temperatures used for the standard calibration.}
}
\value{
dataframe of corrected sample readings in units of microM or mg
}
\description{
function long description
}
\examples{
mydata <- mimsy(filename = "./RawData/MIMS_data.csv", bgcorr = FALSE, std.temp = c(12.5, 13.2))

}
\references{
Garcia, H., and L. Gordon (1992), \emph{Oxygen solubility in seawater: Better fitting
equations}, Limnology and Oceanography, 37(6).

Benson, B. B. & Krause, D. (1984). \emph{The concentration and isotopic
fractionation of oxygen dissolved in freshwater and seawater in equilibrium
with the atmosphere.} Limnology and Oceanography, 29(3), 620-632.
doi:10.4319/lo.1984.29.3.0620

Stull, D. R. (1947). \emph{Vapor Pressure of Pure Substances. Organic and
Inorganic Compounds.} Industrial & Engineering Chemistry, 39(4), 517-540.
doi: 10.1021/ie50448a022

Hamme, R. C. & Emerson, S. R. (2004). \emph{The solubility of neon, nitrogen and argon
in distilled water and seawater}, Deep-Sea Research I, 51(11), 1517-1528.
}
\author{
Michelle Catherine Kelly
}
