% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimsy.R
\name{mimsy}
\alias{mimsy}
\title{\code{mimsy} Calculate dissolved gas concentrations}
\usage{
mimsy(file, baromet.press, units, std.temps, bg.correct = FALSE,
  tz = Sys.timezone(), salinity = 0)
}
\arguments{
\item{file}{the name of the file which the data are to be read from. File must have 'comma seperated value' ('.csv) format and follow the formatting guidelines on the Github page. If the file does not contain an absolute path, the file name is relative to the working directory.}

\item{baromet.press}{the ambient barometric pressure while samples processed on the MIMS. Can be a vector, if more than one reading was taken.}

\item{units}{the units of barometric pressure. Must be one of "atm", "hPa", "psi", "bar", or "Torr".}

\item{std.temps}{a numeric vector (maximum length 2) of the water temperatures (degC) of the standard baths.}

\item{bg.correct}{If `FALSE` (default), no background correction is applied. If `TRUE`, background correction is applied.}

\item{tz}{a character string that specifies which time zone to parse the date with. Defaults to the user's current time zone setting. The string must be a time zone that is recognized by the user's OS.}

\item{salinity}{the salinity of standards, in units of per mille. Defaults to 0.}
}
\value{
a list, containing dataframes of calculated gas concentrations, solubility concentrations, and calibration factors.
}
\description{
Return dissolved gas concentrations in units of micromole and milligram from membrane inlet mass spectrometer (MIMS) signal data
}
\examples{
\dontrun{
dat <- mimsy(file = 'MIMS_data.csv', baromet.press = 981.2, units = 'hPa', std.temps = c(12.5, 15.2))
}

}
\references{
Garcia, H., and L. Gordon (1992), \emph{Oxygen solubility in seawater: Better fitting
equations}, Limnology and Oceanography, 37(6).

Benson, B. B. & Krause, D. (1984). \emph{The concentration and isotopic
fractionation of oxygen dissolved in freshwater and seawater in equilibrium
with the atmosphere.} Limnology and Oceanography, 29(3), 620-632.
doi:10.4319/lo.1984.29.3.0620

Stull, D. R. (1947). \emph{Vapor Pressure of Pure Substances. Organic and
Inorganic Compounds.} Industrial & Engineering Chemistry, 39(4), 517-540.
doi: 10.1021/ie50448a022

Hamme, R. C. & Emerson, S. R. (2004). \emph{The solubility of neon, nitrogen and argon
in distilled water and seawater}, Deep-Sea Research I, 51(11), 1517-1528.
}
